MB-59183 Detect references to function arguments as correlation
If an inline UDF is defined with arguments, then currently for all
statement inside the UDF, we mark the statement with "hasVariables",
which loosely represents that the statement COULD potentially
reference function arguments and thus we should treat it similarly
as correlated and re-execute the statement every time. Note that
this is done for all statements inside the UDF regardless of
whether a statement actually references any function argument or
not.

There are a couple of issues with this approach:

1. performance. If a statement (represented as a subquery inside
   UDF) has no actual references to any of the function arguments,
   we could potentially cache and reuse the subquery result,
   however this is currently disabled for all statements inside UDF.

2. expression scan. The "hasVariables" marking is currently only on
   a subquery. We also have a result cache for expression scan, and
   that only checks for correlation, not "hasVariables". This could
   potentially leads to wrong result (i.e. incorrect caching of
   result when it should not).

The fix introduces a new mechanism to handle function arguments.
When formalizing subquery expression inside UDF, we already add all
function arguments to the WITH of the formalizer as "permenant",
which then enables the formalizer to detect any references to these
function arguments. We now treat such references as correlation
references, thus if a statement inside UDF references any of the
function arguments it'll be marked as correlated. A statement that
has no such references can then take advantage of result cache.

We also modify the expression scan result cache, to not cache the
annotated value, instead cache the underlying value. When using the
result cache, a new annotated value will be created using the
current "parent" value, such that when we use the cache we don't
just get the value of the "parent" from the very first execution.
Similar changes are already done for result cache in IndexScan3 and
Fetch (for result cache when using primary index on inner of a
nested-loop join).

With this change the "hasVariables" mechanism is now obsolete and
removed. We effectively combined the previous correlated and
hasVariables mechanisms into the same correlated mechanism.

We also no longer push the function arguments as variables as
allowed references in the beginning. Similar change is already done
for WITH aliases (MB-44279). Now both with alias and function
arguments use similar mechanisms.

Change-Id: I6381213816c0a39ad28df95c4bed776725289f1d
Reviewed-on: https://review.couchbase.org/c/query/+/199060
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>