MB-59541 Inherit parent query's CTE definition in formalizer
This is a partial backport of MB-44279, it's actually a fairly
small portion of the original fix.

During formalization process the CTE (WITH clause) information
is stored in the formalizer. When we need to formalize a
subquery we save the parent's CTE information, but we currently
only inherit the "permanent" WITH information (which represents
function arguments for inline UDFs), but not regular CTE, as
such when formalizing the subquery, we lose track of the defined
CTEs in the parent query, which can leads to various issues
including incorrect correlation reference determination
(a reference to the CTE in the parent scope should not be
treated as a correlation reference).

The fix now has a marker for subquery in FormalizeSubquery(),
and if we are creating a new formalizer for a subquery we
inherit the parent query's CTE definitions, since these CTEs
can be referenced in the subquery as well, and should not be
treated as correlation references.

Change-Id: I46f1d2ab207e739815ebace4cf7a5233d6ed15e1
Reviewed-on: https://review.couchbase.org/c/query/+/199062
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/200402