MB-59214 Formalization changes for CTE
This is a backport of MB-44279, as well as the regressions caused
by the original fix of MB-44279, including:
MB-55835, MB-56431, and part of MB-57904.

MB-44279 Formalization changes for CTE

Previously if a CTE is used, it is first pushed as a binding
variable in the formalizer such that it is "allowed" as a
reference anywhere in the SELECT query.

Normally for any FROM clause term we add it to the "allowed"
list when it is first used, then it is available for any
subsequent clauses, and if the same alias is used subsequently
we can detect that and raise an error. This mechanism does not
work for a CTE variable since it becomes "allowed" when the
CTE definition is encountered, before any actual references to
the CTE variable. To make subsequent references work we had to
disable the "duplication" checking and thus the same mechanism
for detection of duplicated aliases does not work for CTE
variables.

To fix this issue, we need to change the way CTE variables are
handled in the formalizer, i.e. we also need to only add a CTE
variable to be "allowed" when it first appears in a FROM
clause, this way similar duplicate detection will work for CTE
variable as well.

To make this work, when CTE definition is encountered, we only
put the CTE information in the "withs" field in the formalizer,
which we already do previously, but now we need to keep track
of a few additional information, as a result a new WithInfo
structure is defined and populated as a CTE is being processed.

If a CTE is correlated, instead of relying on the formalizer
itself to keep this info (since the CTE is pushed to the
formalizer early on), we now keep this info in the new WithInfo
structure itself, and transfer the correlation info as needed
when the CTE is referenced.

We also support referencing a CTE as a general expression, i.e.,
the CTE does not need to be referenced in the FROM clause first
before it can be references as an expression. To make this work
when an identifier is formalized we need to check and handle if
the identifier is a CTE reference.

MB-55835 Proper determination of coveirng index scan with CTE references

Prior to fix of MB-44279 CTE variables are marked as "binding
variables"; with the fix for MB-44279 CTE variables are now
marked as "with aliases". The current code for checking
covering of an identifier thus needs to be updated to handle
"with aliases" in addition to "binding variabbles".

MB-56431 Detect conflicts with WITH aliases during formalization

Prior to the fix for MB-44279, WITH aliases are pushed to
the formalizer in the beginning of a SELECT statement as
a binding variable. This prevents any other usage of the
WITH alias in the query, e.g. an alias used in the FROM
clause, or as a binding variable name. One issue with
this approach is that we cannot effectively check the
usage of an WITH alias as an ExpressionTerm (FROM clause)
as indicated by MB-44279.

The fix for MB-44279 changed the formalization for WITH
aliases. We now treat it similar as any other FROM term
alias, i.e. if the WITH alias is used in the FROM term,
at that time we push it as an allowed alias, otherwise,
any other references to the WITH alias will treat it as
a general expression reference.

One side effect of the fix for MB-44279 is that we are
no longer checking for duplicated references to the WITH
alias, i.e. we now allow for example the usage of a WITH
alias as a alias for a FROM term (keyspace, expression,
subquery), or as a binding variable. Although one can
argue that this behavior is acceptable, which effectively
means that we are overriding the WITH alias with the new
usage, this does create potential confusion, and is a
different behavior than previous releases.

Therefore we decide to add back the check for duplicated
WITH alias references, the fix includes:
 - add a new member function AllowedAlias() to formalizer
   that checks for allowed alias, with two extra boolean
   arguments:
   - checkWith: whether to include WITH alias in check
   - curScope: whether current scope only (i.e. exclude
     parent scope)
 - add a new member function HasAlias() to formalizer
   that checks for existing alias
 - remove current member functions Allowed() and Aliases()
   from formalizer, caller must use one of the above new
   member functions
 - in situations where we want to include WITH alias in
   duplication check, e.g. FROM clause aliases, use
   AllowAlias() checkWith = true and curScope = false,
   this will be majority of situations
 - in situations where we want to exclude WITH alias in
   duplication check, e.g. when handling ExpressionTerm
   where WITH alias is checked separately, use AllowAlias()
   with checkWith = false and curScope = false
 - in situations where we want to check against the
   current scope only (and exclude WITH alias), e.g.
   when checking the FOR clause in index join/nest
   (JOIN <term> ON KEY <key> FOR <keyspace>), use
   AllowAlias() with checkWith = false and curScope = true
 - check for existence of an alias is currently only
   used in ExpressionTerm, and now it uses HasAlias()
 - check for binding variable is handled specially since
   it is in the formalizer code itself, where we just
   add an additional check for WITH aliases
 - fix an existing bug with handling of set operations
   (UNION/EXCEPT/INTERSECT) where we incorrectly set
   projection term alias to be a keyspace alias

MB-57904 Proper propagation of WITH clause correlation

1. When a WITH clause contains correlation, we mark the WITH
   clause itself as correlated, but we did not put the correlation
   references properly, such that if this WITH clause is within a
   subquery the correlated information is not properly propagated
   up.

The fix contains:

1. Add logic to add correlation references from WITH clause to the
   formalizer itself such that it can be detected further up..

Change-Id: Ia5d79f23e97fc961c3631f34c19a1f570a3069ca
Reviewed-on: https://review.couchbase.org/c/query/+/189770
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/199130
Well-Formed: Restriction Checker