MB-59544 Allow hash join to reopen
This backport includes part of the changes for MB-54197, which
adds a few functions to the SimpleFromTerm interface for lateral
join. This is required for the backport of MB-59082.

The reopen logic is currently not implemented for hash join, since
we don't want the hash table to be repeatedly built. However in
some cases we need to repeated execute a part of the query plan,
e.g. a correlated subquery, or a statement inside an inline UDF.

We now allow hash join to be reopened. The hash table is still
scoped to one execution, i.e. built in beforeItems() and dropped
in afterItems(). There is no preserve/reuse of the hash table in
this checkin.

With this capability we can also now support hash join in case
of correlation. We still check and prevent hash join from being
used in case of lateral correlation.

Change-Id: Ic92e4e81f55502f323fc56264fa0faa561ddec42
Reviewed-on: https://review.couchbase.org/c/query/+/199123
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/199183
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/200416