MB-59240 Match index already exist error better
In the errors package we attempt to match certain message text
returned from GSI and use a more specific error code. Specifically
for "index already exists" error we currently match the pattern:
"Index (.*) already exists"
However I noticed during migration testing that there are times
that GSI returns the error that just says "Index already exist.",
i.e. it uses "exist" instead of "exists". This causes the pattern
matching to fail and thus not catching this as an index exists
error.

We now remove the trailing 's' in the pattern matching such that
it catches the error message text with "exist" in addition to
"exists".

Change-Id: I5a6c1294f4bce9a55377155cabf3378231b994d6
Reviewed-on: https://review.couchbase.org/c/query/+/199213
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>