MB-57717 Do not wait for avaialibity of primary index before staring migration
Currently for CBO stats migration, we wait for the system collection to be
available, as well as the primary index on the system collection to be
available before starting migration on that bucket. There is no need to
wait for the primary index since the migration process does not use it.
Waiting for primary index to be available just delays the migration process
unnecessarily. The UDF migration does not wait for the availability of the
primary index on the system collection.

Modify the logic for CBO stats migration such that it does not wait for
the primary index before starting migration. Once migration is complete
for a bucket (for both UDF and CBO stats migrations) we fire off a background
process to build the primary index on the system collection for that bucket.

Change-Id: I2a199debda40868f92f3d063580d8e5d60b16079
Reviewed-on: https://review.couchbase.org/c/query/+/199257
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>