MB-57717 Consolidate migration retry logic
There are two paths for migration, one if all buckets already
have system collection available at start up time, which goes
through migrateAll(), otherwise it goes through bucket-by-bucket
triggered through the bucket updater for each bucket. The retry
logic currently is also separate for the two paths.

The retry logic for migrateAll() is not very efficient since
it does not keep track of the migration state for each bucket,
it simply goes through all buckets each time regardless of
whether the bucket has been migrated or not.

Consolidate the retry logic such that for migrateAll() we goes
through all buckets the first time, via regular mechanism, but
for retry logic we go through the same logic for the other path,
such that we have a single retry logic for migration.

Also change the wait logic to use a sync.Cond when waiting for
migration to start.

Change-Id: I50c854dc699e768fb24ad3013dd31958c8145c3f
Reviewed-on: https://review.couchbase.org/c/query/+/199288
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>