MB-57717 Do not scan system collection while migration is ongoing
We have a mechanism that checks for scopes that are dropped outside
of query and try to clean up the entries in the system collection
for such scopes. This is done when a bucket is being loaded.

During migration we also need to load a bucket (multiple times) and
this check is not needed, partly because migration is done at node
start up time, partly because the system collection is in the
process of being populated.

Add a mechanism to mark migration state in datastore (in addition
to metakv) such that we can determine in datastore whether migration
is complete or not, and avoid trying to cleanup obsolete entries
from the system collection during migration process.

Change-Id: Ibadfda368b64be1f4789737a0771bcb319cdaa49
Reviewed-on: https://review.couchbase.org/c/query/+/199453
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>