MB-57717 Add check for local migrations info
Currently the check for migration complete is to scan the old
location (metakv for UDF and N1QL_SYSTEM_BUCKET for CBO stats)
and make sure it does not contain any remaining entries.

In case there is an issue with the cluster (e.g. one of the
nodes died) we may have undetected issues with the scan and
incorrectly assume all entries are done, and prematurelly
conclude that migration is done.

Add an extra check for the locally gathered migration state
as well, just to be sure.

Change-Id: I59b2922038ee719c6737965d5934b97fa37e80ad
Reviewed-on: https://review.couchbase.org/c/query/+/199595
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>