MB-57717 Regard migration as complete if no migration is needed
When a query node starts up, if it detects that migration is not needed,
e.g. it has been done previously or it is done by a different node and
marked in metakv as complete, then no migration infrastructure will be
allocated in datastore. In such cases we need to return true from
datastore.IsMigrationComplete() function.

Change-Id: If6164b79ef8a22da223af0568982f9d6fedd0bd9
Reviewed-on: https://review.couchbase.org/c/query/+/199685
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>