MB-59369 Add ability to detect internal index scan errors on system collection
For accessing information stored in system collection (_system._query)
under each bucket, we use an index scan (primary index or sequential scan).
In case there is an error from the index scan, we currently do not track
such errors and simply return no result. As such, the caller does not know
whether it gets no result because there is nothing in the underlying
system collection, or there is an error from the index scan.

For CBO stats, we use similar mechanism to access the old CBO stats
location, in N1QL_SYSTEM_BUCKET.N1QL_SYSTEM_SCOPE.N1QL_CBO_STATS. This
could be an issue during migration when we are trying to move CBO stats
records from the old location to the new location. In case there is an
error during index scan of the old location, we could incorrectly
determine that there is no CBO stats to be migrated.

The internal index scan currently uses datastore.NULL_CONTEXT, which
does not do anything in case of error. To fix this, we add a new context,
datastore.SYSTEM_CONTEXT, which is built on top of NULL_CONTEXT, except
it keeps track of errors, and can return errors when requested. Then
during internal index scan on system collection, check and return errors
if exist.

Change-Id: I68b5c0ec7c0e7fd96ad1a1657d6b87c1f208b614
Reviewed-on: https://review.couchbase.org/c/query/+/199791
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>