MB-57717 Better handle unavailability of N1QL_SYSTEM_BUCKET
During CBO stats migration we need to scan N1QL_SYSTEM_BUCKET to
get the old entries as well as determining whether there are any
entries left (migration complete). The scan is done using the
primary index.

If there is an issue in accessing N1QL_SYSTEM_BUCKET during
migration, e.g. if one of the nodes goes down which renders the
bucket inaccessible, we may get a nil pointer when trying to
obtain a handle for N1QL_SYSTEM_BUCKET, in which case we may
incorrectly conclude that we don't have any CBO stats to be
migrated, and drop N1QL_SYSTEM_BUCKET before the stats can be
migrated.

To better handle potential unavailability of N1QL_SYSTEM_BUCKET,
the following changes are made:

- Check availability of N1QL_SYSTEM_BUCKET at the beginning of
  migration, and set a variable indicating whether it's present
- During migration, if we get a nil handle for N1QL_SYSTEM_BUCKET,
  but we know it is present due to the earlier indicator, treat
  it as error condition
- Add an indicator to the scan function for N1QL_SYSTEM_BUCKET,
  such that in case it gets a nil handle it will return an error
  in case we expect a valid handle
- The check for migration complete is only needed if the
  migration of the current bucket is successful (complete)

Change-Id: I47a0bac8a9412d499108eaf4847062c933930e2a
Reviewed-on: https://review.couchbase.org/c/query/+/199866
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>