MB-30815 Add the key as a document field in INFER ...
... sampled documents when the flag INCLUDE_KEY is specified.
Implicitly specify this flag for INFER on keyspaces when no other flags
have been specified.  (Can be included with other flags explicitly if so
desired.)  This means that expression INFER, CBO etc. are unaffected but
the UI gets the benefit with no changes.

Adding the object field "~meta" with a single field "id" allows the keys
to be sampled and present with examples in the UI.  ("~meta" follows
other implicit naming and attempts to ensure it is after all user fields
in the default order.)

Change-Id: If65f02d1b72e23a6bc42f7024073c94ffab8b08d
Reviewed-on: https://review.couchbase.org/c/query/+/200232
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>