MB-59545 Move Park & Resume to separate interface
With Park & Resume part of expression.Context we lose the ability to
cast from execution.Context where needed for expression evaluation.  As
a separate interface the UDFs can retain the ability to deal with
parking and resuming whilst preserving the inferencer functionality as
is.

MB-59545 Use operator context for Advise statement

The Advise statement requires an operator context. For an
Advise statement issued directly, this is already handled
in Server.getPrepared() function. However, for an internally
generated Advise statement, e.g. from an Advisor() function,
we still use the original execution context, which could
leads to runtime issues.

Fix the internally geneated Advise statement to use operator
context, in function PrepareStatement(), the same way as
an Advise statement issued directly.

Change-Id: I2ea31af77eaf254d727fb4fc83b36fb713b50474
Reviewed-on: https://review.couchbase.org/c/query/+/184566
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/200097
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/200406