MB-59540 [7.2.3-MP1]: Do not use cached expression scan if potential references to function arguments exist
This is a different fix for MB-59183 than what is in 7.2.4/trinity/master,
since in 7.2.4 and above the fix for MB-59183 requires the fix for
MB-59214/MB-44279, which is a complete rework of how WITH clauses are
handled in the server. We do not want to put that complicated change
in 7.2.3-MP1, instead, we fix it in a more conservative way.

For ExpressionScan, add an indicator of whether it is under a UDF
with arguments, and if so, do not use cached results. This is
similar to how we currently (in 7.2.3) handles cached subquery result.

Change-Id: Ifa0d26402b89f27312d71c2bfc603e3194a3d288
Reviewed-on: https://review.couchbase.org/c/query/+/200420
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/200631