MB-59552 Do not use join predicates as index span if under hash join
When generating an index span for an index scan under hash join, we
should not use join predicates since we cannot evaluate expressions
from the other keyspace yet. This behavior is correct in 6.5 and 6.6,
however, in 7.0 and above, with the fix for MB-30813, in order to
allow correlation variables to be used in an index span, we added
an additional check for whether the query block is correlated. The
issue with that is the check does not properly distinguish between
a correlation variable and a reference to a different keyspace in
the same query block.

The fix is to add a check in case of correlated subuqery, to see
whether the expression references any other keyspaces in the same
query block.

MB-59552 Do not consider primary index for index scan inside correlated subquery

We don't allow primary index scans for a keyspace inside a correlated
subquery (for performance reasons). Current code would add primary
index to the list of indexes to be considered, and it is possible
that when CBO is on we could prefer a primary index due to lower
cost. Later on we detect primary index and return an error, even
though a usable secondary index exists.

The fix is to exclude a primary index for a keyspace inside a
correlated subquery.

Change-Id: Ie5d9bccc9a8c118b325a3837d3de08660347cb45
Reviewed-on: https://review.couchbase.org/c/query/+/183112
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/200426
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>