Clone master in case a branch may not exist
When building a branch (e.g. trinity) and we need to clone from
a different repo (e.g. sigar), that repo may not have the same
branch and thus we need to clone from master first, and let
subsequent logic (checkout_if_necessary) to attempt to get it
to the proper branch (if exists).

Change-Id: Iade819ec0c3275301a9ba99f0a47f5766c11675a
Reviewed-on: https://review.couchbase.org/c/query/+/200746
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>