MB-59642 Replace LET expressions when performing covering determination
When a LET clause is used in a query, we "dereference" the LET bindings
in the WHERE clause, i.e., replace references of any LET variable with
its actual binding expression, before performing index selection.
However, we do not do such "dereference" any other place currently,
specifically, when we are doing covering determinations, we use the
original expressions (WITH let variable references).

This is normally not an issue, since the expressions do include the
LET binding expressions and thus when we do covering determinations
the LET bindings are considered. In this bug though, the LET variable
binds to an object, and then we further reference subfields of the
object. This results in a mismatch at the expression level, since
the original expression has v.field (where v is the LET variable),
but the index key is o.field (where o is the object that v binds to).
This mismatch subsequently results in a missing index key in the
index projection list.

Since we do index selection with the "dereferenced" expressions,
conceptually we should also do the covering determination with
similar "dereferenced" expressions instead of the original
expressions. Add infrastructure to perform such LET binding
dereferences when we need to gather "this.cover.Expressions()",
such that we have a consistent expression matching. A new utility
function, getExprsToCover() is used to replace current usages of
"this.cover.Expressions()".

Change-Id: I36d0516d3ed34ff476e87c8c10c1731aab6e11a5
Reviewed-on: https://review.couchbase.org/c/query/+/201045
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>