MB-59642 Include WITH aliases in gathered aliases map
During planning we gather a map of all aliases used in the query
block, such that when we need to do binding variable replacement
we do not accidentally use an existing alias. The gathered aliases
map currently only contains keyspace aliases and LET binding
variables, since this is done at the Subselect level, and we moved
WITH clauses to the Select level.

Extend the logic such that we also include WITH aliases in the
gathered aliases map.

Change-Id: I875a336b7bf606da8f3f46b533c4b71b9c3aed7e
Reviewed-on: https://review.couchbase.org/c/query/+/201049
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>