MB-59084 Modify condition for filtering on cached result
In trinity we have a couple of enhancements that allow a primary
index (or sequential scan) to be used:
 - inner of nested-loop join
 - in correlated subquery
in both cases a result cache is used on either IndexScan3 (covering)
or Fetch (non-covering). We currently have restrictions on building
additional filters in case of cached result, i.e. if we use cached
result we should not do additional filtering, to avoid messing up
the result cache.

The previous fix for MB-59084 (incorrectly) removed the check for
correlated subquery for filter generation. We need to add that
check back for the covering case such that we do not perform
filtering on cached result in the covering case. This is necessary
since in this case the extra filter is on IndexScan3 itself.

For the non-covering case though, the result cache is on the Fetch
operator, but any extra filter will be on a separate Filter operator
after the Fetch. In this case the extra filter will not mess up the
result cache on the Fetch operator, and thus we can actually allow
extra filter to be done.

Change-Id: I1ac4927d830d79802178d1153906e48445cdf875
Reviewed-on: https://review.couchbase.org/c/query/+/201063
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>