MB-59730 Add optimizer hints info in prepared statement
When a statement is prepared, if the statement contains optimizer
hints, information about the optimizer hints are not included in
the resulting prepared statement. This makes it difficult to
determine whether any of the optimizer hints specified are
followed or not. This makes it more difficult to debug any
potential issues with query plan for a prepared statement.

Add optimizer hints info in a prepared statement, such information
is the same as what's already included in explain, i.e. it lists
the optimizer hints grouped by status, followed, not followed,
error, invalid, unknown. For hints with error, it also lists an
error message.

Also add the same information in output of system:prepareds.

Change-Id: Idcccba4673d680ea99ffb80c87d06dff91c191e6
Reviewed-on: https://review.couchbase.org/c/query/+/201341
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>