MB-59757 Handle disabled SEQ_SCAN
The previous logic does not properly handle the case when
SEQ_SCAN is disabled, i.e. we'll still return an error in
that case.

Modify the fix such that we try primary index first, then
sequential scan, then if nothing works, wait for the creation
of primary index synchronously and try with primary index
again.

Change-Id: Ie1941306ebb1304084cf4cd29fd12edaa0f6babb
Reviewed-on: https://review.couchbase.org/c/query/+/201385
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>