MB-59752: [7.2.4]  Revise CURL() URL related processing and checks
* Store the node's allowList as a list of URL objects instead of
strings This will make future comparison against CURL() input URLs
easier.

* Resolve references, validate protocol, clean up the URLs for CURL().
Perform this transoformation on the input URL in CURL() and
the allowed and disallowed URLs list in the node.

* When checking if the URL is in the Allowed or Disallowed lists -
do not solely rely on prefix matching as this is insufficient

Instead compare each component of the input URL with the URLs
in the lists.

MB-59755:[7.2.4]Revise check in CURL() to detect /diag/eval in url path

The issue with the current check was that although syntactically
multiple continuous forward slashes are allowed as path separators
in a URL path - this was not being checked for.

With the fix for MB-59752 - this issue gets fixed. Since now perform
the diag/eval detection on the tranformed URL where all instances
of multiple adjacent forward slashes are replaced with a single
forward slash.

Change-Id: Ib8abba8977fb67a3c962fc109cd1ebc70e0845d9
Reviewed-on: https://review.couchbase.org/c/query/+/201441
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>