MB-59847 Handle index group aggregates pushdown properly in CBO
When CBO is in use, if there are multiple indexes capable of
index group aggregates pushdown (full or partial), currently
CBO always choose the index with the "maximum" pushdown, i.e.
an index with full group aggregates pushdown is always favored
over an index with partial group aggregates pushdown,
regardless of cost. This may in some instances pick the wrong
index.

Instead, we now try to add an extra cost to account for
"future" group/aggregation to an index that does not have
full group/aggregates pushdown. Similar for order or partial
order pushdown. This is not exact but is the closest we can
do in order to use cost to compare two indexes, instead of
always use the index pushdown property.

Also add an function to compare index pushdown properties,
to treat full/partial group/aggregates/order pushdown as
"similar" to each other, and attempt to use cost to pick
the better index. This is used in case we cannot get
proper group/sort cost. This allows us to still use index
pushdown property somewhat, i.e. an index with group/aggregate
pushdown is favored over one without it. Again this is only
used when we cannot get proper group/sort cost.

Change-Id: I0e7632f5f6e266f09dbff8c1c74576b9f89fcb7c
Reviewed-on: https://review.couchbase.org/c/query/+/201730
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/201799
Well-Formed: Restriction Checker