MB-59885: Revise handling of Get and Request CURL() options
Aligns behaviour of handling "Get" and "Request" CURL() options within
reason to the previous go-curl based implementation.

This fix is to address handling conflicting values of these two
options.

Details:
If the "Get" option is set and is True - the request is a GET request.
But if it is Unset or False and the "Request" option is set -
the request is whatever method set in the "Request" option.

But if the "Request" option is Unset, use default Curl behaviour. i.e
request is a GET unless "data" options are present - then it is a POST
request.

Change-Id: I40fa6a6b35def50d0722f45570b5456082f27883
Reviewed-on: https://review.couchbase.org/c/query/+/201839
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>