MB-59885: [7.2.4] Revise handling of Get and Request CURL() options
Aligns behaviour of handling "Get" and "Request" CURL() options within
reason to the previous go-curl based implementation.

This fix is to address handling conflicting values of these two
options.

Details:
If the "Get" option is set and is True - the request is a GET request.
But if it is Unset or False and the "Request" option is set -
the request is whatever method set in the "Request" option.

But if the "Request" option is Unset, use default Curl behaviour. i.e
request is a GET unless "data" options are present - then it is a POST
request.

Change-Id: I139b59b3ff6081a7b5ee6e2237b5da0661fce147
Reviewed-on: https://review.couchbase.org/c/query/+/201840
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker