MB-60011 Allow covered scan on primary index inside correlated subquery
We do not allow using primary scan if it is inside a correlated subquery.
However, if a covering index scan can be used with a primary index,
previously we allow such plans since a covering index scan with primary
index is not considered to be a primary scan.

Such usage is no longer permitted after the fix for
MB-53565/MB-54045/MB-59552 (in Trinity, 7.2.0 and 7.1.6-MP1, respectively),
since we excluded primary index from consideration altogether.

The fix is to change the way we prevent primary scan usage inside a
correlated subquery. Instead of not adding the primary index at all, we
still add primary index to the index list, and only when we are
considering secondary index scans we skip a primary index if it is
inside a correlated subquery.

Change-Id: I5666312ad43ece06d8e1c53fab85220ed016e3c3
Reviewed-on: https://review.couchbase.org/c/query/+/202226
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>