MB-60011 Allow primary index with span inside correlated subquery
Modification of previous fix. It turns out that previously we also allow
a primary index to be used inside a correlated subquery if it is not
covering but there are index spans (there are predicates on meta().id),
in which case we also generate IndexScan3 instead of PrimaryScan3, and
it is not considered to be a primary scan.

Modify the fix such that we check for primary index with no sargable
keys inside correlated subquery, when we compare indexes in function
minimalIndexes(), such that we apply the logic to address the original
concern, which is favoring primary index over secondary index by cost
when CBO is in use. The logic is applied to both CBO and non-CBO code
paths.

Change-Id: Ie80c9f5e883b0442b9f1b5f76747418c7d37791d
Reviewed-on: https://review.couchbase.org/c/query/+/202339
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>