MB-52176 Avoid gathering optimizer statistics when index already exists
In Trinity we added the ability to automatically gather optimizer statistics
when a (non-deferred) CREATE INDEX/CREATE PRIMARY INDEX statement is run,
if CBO is turned on.

If the index already exists and 'IF NOT EXISTS' option is specified, then
we ignore the error for "index already exists", in which case we should not
attempt to gather optimizer statistics.

Change-Id: I926ee6edadd16b1b3e027dd6dfd6b2ad896fae79
Reviewed-on: https://review.couchbase.org/c/query/+/202384
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>