MB-60260: Fix incorrect COUNT(*) returned on system:indexes
..when the CountScan Operator performs the count i.e no predicate

SELECT COUNT(*) FROM system:indexes;

The above query returns a higher count than the number of entries
present in system:indexes.

This change fixes the following in Count on system:indexes:

i. When an index is defined on a bucket - system:indexes does not
   contain entries for both the instance on the bucket and the
   bucket's default._default collection.

   Yet both of these index instances were counted. Which is incorrect.

ii. System:indexes does not include entries for indexes created on
    system collections i.e collection names prefixed with an underscore

    Yet these entries were counted. Which is incorrect.

To note-
This only changes Count() handling for system:indexes.
The behaviour of Count() on system:all_indexes remains the same

Change-Id: I22b237e5f3d1d40d02afbef0b67dd00482c70ca6
Reviewed-on: https://review.couchbase.org/c/query/+/203292
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>