MB-60360 Reduce size of annotatedValue
META has a fixed list of well defined fields.  Replacing a
map[string]interface{} with a simple structure saves a noticeable
portion of memory.  This change naturally requires an API change to
interact with it.  Mimicking the existing get/set for fields rather than
individual get/set methods for structure members reduces the overall API
changes. (The API changes mean meta logic is contained within the
annotated value.)

Most attachments are well-known names too, so these at least should be
constants rather than literals where used.  Replacing the
map[string]interface{} with a map[int16]interface{} still leaves us with
a reasonable number of attachments (I'd argue map[byte]interface{}
would be enough and would be better still) and saves us a lot too.

The special case though is that we do - for WINDOW and ORDER BY - label
attachments with non-well-known values.  To accommodate this a well
known attachment is added with an array of strings identifying the
following indices.  Lookup is a linear search through this array: in
most cases it will not be that long & we're trying to avoid memory
wasting maps.  As documents may have differing sets of attachments prior
to arriving at a WINDOW or ORDER BY, we can't pre-compute this list.

Change-Id: Id74a01dc12baf1efd41d2320e1d33530b3662df7
Reviewed-on: https://review.couchbase.org/c/query/+/206489
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>