MB-60291 Remove permissions based plannning...
... for sequential scans.  The original idea was to completely exclude
possible plans with sequential scans when the preparing user lacked the
necessary privileges for them, resulting in the same error at prepare
time that would have been seen in prior versions.  But this doesn't
translate for re-prepare operations.  Within a same-version cluster the
plans could be distributed without re-preparing so worked; upgrade
exposes the issue as distributed plans have to be re-prepared owing to
version differences.

With this change when an attempt is made to prepare a statement where a
sequential scan is the chosesn access method and the user lacks the
necessary privilege, a privilege error (13014) is raised.

Change-Id: I4aa58f7181bbcdea27fadce1c9f119dcd3e17956
Reviewed-on: https://review.couchbase.org/c/query/+/203354
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>