MB-60294: [7.2.5] Fix incorrect COUNT(*) returned on system:indexes
..when the CountScan Operator performs the count i.e no predicate

SELECT COUNT(*) FROM system:indexes;

The above query returns a higher count than the number of entries
present in system:indexes.

This change fixes the following in Count on system:indexes:

i. When an index is defined on a bucket - system:indexes does not
   contain entries for both the instance on the bucket and the
   bucket's default._default collection.

   Yet both of these index instances were counted. Which is incorrect.

To note-
This only changes Count() handling for system:indexes.
The behaviour of Count() on system:all_indexes remains the same

Change-Id: Iee534ed808218050a1ad846a82d019b8a56580d7
Reviewed-on: https://review.couchbase.org/c/query/+/203410
Well-Formed: Restriction Checker
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>