MB-60410 Avoid exponential checking for redundancy
The fix for MB-56077 has a new way of checking and avoid adding
redundant filters as index filters in case of OR clause. This
causes performance issues for this bug where we have an OR clause
of 2 subclauses, but each one has a long list of OR clauses ANDed
together. The current check uses SubsetOf() function which in case
of AND/OR clause walks the children and recursively calls itself.
The number of comparisons thus increase quickly when both sides of
comparison has a large number of children.

The fix addresses this issue in a couple of aspects:

 - we already track whether we've modified the expression, and in
   case the expression has not been modified at all so far, we can
   avoid the check (just use the original expression in this case)
 - if the check is necessary, only if the term being checked is
   not an AND expression we use SubsetOf(), otherwise do a simpler
   EquivalentTo() call instead, which is much more efficient.

Change-Id: I268b1eb950c2cd17814a716bca2a384314e87c66
Reviewed-on: https://review.couchbase.org/c/query/+/203970
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>